+++ app/src/main/java/com/example/todo/airbnb/presentation/search/date/DateScreen.kt
@@ -1,4 +1,4 @@
-package com.example.todo.airbnb.presentation.search.components.searchmap
+package com.example.todo.airbnb.presentation.search.date
 
 import androidx.compose.material.Button
 import androidx.compose.material.Text
@@ -7,8 +7,11 @@ import androidx.navigation.NavController
 import com.example.todo.airbnb.presentation.main.components.Destinations
 
 @Composable
-fun SearchMapScreen(navController: NavController) {
-    Button(onClick = { navController.navigate(Destinations.searchCondition) }) {
-        Text(text = "SearchMap")
+fun DateScreen(navController: NavController) {
+    Button(
+        onClick = { navController.navigate(Destinations.fare) }
+    ) {
+        Text(text = "Calender")
     }
+
 }
