+++ app/src/main/java/com/example/todo/airbnb/presentation/search/detail/DetailScreen.kt
@@ -1,9 +1,8 @@
-package com.example.todo.airbnb.presentation.search.components.detail
+package com.example.todo.airbnb.presentation.search.detail
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.example.todo.airbnb.data.Travel
 
 @Composable
 fun DetailScreen(navController: NavController) {
